fix: fix the remaining tests
