+++ tests/unit/bff/v1/routers/test_img2img.py
@@ -1,29 +1,9 @@
-import base64
-import os
 from typing import Callable
 
 import pytest
 import requests
-from docarray import Document, DocumentArray
+from docarray import DocumentArray
 from starlette import status
-
-
-@pytest.fixture
-def base64_image_string(resources_folder_path: str) -> str:
-    with open(
-        os.path.join(resources_folder_path, 'image', '5109112832.jpg'), 'rb'
-    ) as f:
-        binary = f.read()
-        img_string = base64.b64encode(binary).decode('utf-8')
-    return img_string
-
-
-@pytest.fixture
-def sample_search_response_image() -> DocumentArray:
-    result = DocumentArray([Document()])
-    matches = DocumentArray([Document(uri='match')])
-    result[0].matches = matches
-    return result
 
 
 def test_image_index_fails_with_no_flow_running(
