+++ tests/unit/bff/v1/routers/test_txt2img.py
@@ -2,24 +2,21 @@ from typing import Callable
 
 import pytest
 import requests
-from docarray import Document, DocumentArray
+from docarray import DocumentArray
 from starlette import status
 
 
-@pytest.fixture
-def sample_search_response_text() -> DocumentArray:
-    result = DocumentArray([Document()])
-    matches = DocumentArray([Document(text='match')])
-    result[0].matches = matches
-    return result
-
-
-def test_text_index_fails_with_no_flow_running(client: requests.Session):
+def test_image_index_fails_with_no_flow_running(
+    client: requests.Session, base64_image_string: str
+):
     with pytest.raises(ConnectionError):
-        client.post(f'/api/v1/text-to-image/index', json={'texts': ['Hello']})
+        client.post(
+            '/api/v1/text-to-image/index',
+            json={'images': [base64_image_string]},
+        )
 
 
-def test_text_search_fails_with_no_flow_running(client: requests.Session):
+def test_image_search_fails_with_no_flow_running(client: requests.Session):
     with pytest.raises(ConnectionError):
         client.post(
             f'/api/v1/text-to-image/search',
@@ -27,16 +24,7 @@ def test_text_search_fails_with_no_flow_running(client: requests.Session):
         )
 
 
-def test_text_search_fails_with_incorrect_query(client):
-    response = client.post(
-        f'/api/v1/text-to-image/search',
-        json={'image': 'hello'},
-    )
-    assert response.status_code == 500
-    assert 'Not a correct encoded query' in response.text
-
-
-def test_text_search_fails_with_emtpy_query(client: requests.Session):
+def test_image_search_fails_with_empty_query(client: requests.Session):
     with pytest.raises(ValueError):
         client.post(
             f'/api/v1/text-to-image/search',
@@ -44,20 +32,21 @@ def test_text_search_fails_with_emtpy_query(client: requests.Session):
         )
 
 
-def test_text_index(
+def test_image_index(
     client_with_mocked_jina_client: Callable[[DocumentArray], requests.Session],
+    base64_image_string: str,
 ):
     response = client_with_mocked_jina_client(DocumentArray()).post(
-        '/api/v1/text-to-image/index', json={'texts': ['Hello']}
+        '/api/v1/text-to-image/index', json={'images': [base64_image_string]}
     )
     assert response.status_code == status.HTTP_200_OK
 
 
-def test_text_search_calls_flow(
+def test_image_search_calls_flow(
     client_with_mocked_jina_client: Callable[[DocumentArray], requests.Session],
-    sample_search_response_text: DocumentArray,
+    sample_search_response_image: DocumentArray,
 ):
-    response = client_with_mocked_jina_client(sample_search_response_text).post(
+    response = client_with_mocked_jina_client(sample_search_response_image).post(
         '/api/v1/text-to-image/search', json={'text': 'Hello'}
     )
 
@@ -68,15 +57,15 @@ def test_text_search_calls_flow(
     assert results[0].tags['parameters']['limit'] == 10
 
 
-def test_text_search_parse_response(
+def test_image_search_parse_response(
     client_with_mocked_jina_client: Callable[[DocumentArray], requests.Session],
-    sample_search_response_text: DocumentArray,
+    sample_search_response_image: DocumentArray,
 ):
-    response = client_with_mocked_jina_client(sample_search_response_text).post(
+    response = client_with_mocked_jina_client(sample_search_response_image).post(
         '/api/v1/text-to-image/search', json={'text': 'Hello'}
     )
 
     assert response.status_code == status.HTTP_200_OK
     results = DocumentArray.from_json(response.content)
-    assert len(results) == len(sample_search_response_text)
-    assert results[0].text == sample_search_response_text[0].text
+    assert len(results) == len(sample_search_response_image[0].matches)
+    assert results[0].uri == sample_search_response_image[0].matches[0].uri
